backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
requests>=2.32
python-dotenv>=1.0
chromadb>=0.5.0
openai>=1.30.0